Create app.py
Hey Hanna, at this point we can use the app.py as the processor and import the sentiment analyzer. 
I have made it to the point when it takes the URL input and returns it without crashing the app. Now we have to work on passing that to the sentiment analyzer module and run the sentiment and input the result, let me know if you make some progress.